LayoutTests/media/track/audio/audio-track-mkv-vorbis-addtrack-expected.txt

Testing audio track "addtrack" events.

EVENT(addtrack)
EXPECTED (event.track == 'video.audioTracks[0]') OK

EVENT(addtrack)
EXPECTED (event.track == 'video.audioTracks[1]') OK

EVENT(canplaythrough)
END OF TEST
